refactor: renamed composables and deleted unused
