+++ components/module/ContractTransactions.vue
@@ -13,7 +13,7 @@ import { useContractTransactions } from '~/composables/useContractTransactions'
 import { useFormat } from '~/composables/useFormat'
 import { useSharedData } from '~/composables/useSharedData'
 import { useScreenSize } from '~/composables/useScreenSize'
-import { exportableToCsv, downloadCSV } from '~/composables/csv'
+import { exportableToCsv, downloadCSV } from '~/composables/useCSV'
 import { useBlocks } from '~/composables/useBlocks'
 
 defineOptions({ name: 'ContractTransactions' })
@@ -114,7 +114,7 @@ const subtitle = computed(() => {
   const oldestTxIndex = Math.max(newestTxIndex - pageCount + 1, 1)
   const formattedNewest = new Intl.NumberFormat().format(newestTxIndex)
   const formattedOldest = new Intl.NumberFormat().format(oldestTxIndex)
-  return `(Showing transactions between #${formattedOldest} to #${formattedNewest})`
+  return `Showing transactions between #${formattedOldest} to #${formattedNewest}`
 })
 
 const { transactionStatus } = useStatus(lastBlockHeight)
@@ -286,7 +286,7 @@ function downloadData() {
           class="flex items-center gap-2 px-2 py-1 text-[12px] font-normal text-[#f5f5f5] bg-[#151515] border border-[#222222] rounded-md hover:bg-[#252525] whitespace-nowrap"
         >
           <IconDownload class="w-4 h-4 text-[#bbbbbb]" />
-          <span class="hidden md:inline">Download Page Data</span>
+          <span class="hidden md:inline">Download</span>
         </button>
       </template>
