+++ components/token/TokenTransfers.vue
@@ -8,7 +8,7 @@ import SkeletonTable from '~/components/skeleton/Table.vue'
 import FilterSelect from '~/components/FilterSelect.vue'
 import Tooltip from '~/components/Tooltip.vue'
 import Copy from '~/components/Copy.vue'
-import { exportableToCsv, downloadCSV } from '~/composables/csv'
+import { exportableToCsv, downloadCSV } from '~/composables/useCSV'
 import { useTokenTransfers } from '~/composables/useTokenTransfers'
 import { useSharedData } from '~/composables/useSharedData'
 import { useFormat } from '~/composables/useFormat'
@@ -83,7 +83,7 @@ const subtitle = computed(() => {
   const oldestIndex = Math.max(newestIndex - pageCount + 1, 1)
   const formattedNewest = new Intl.NumberFormat().format(newestIndex)
   const formattedOldest = new Intl.NumberFormat().format(oldestIndex)
-  return `(Showing transfers between #${formattedOldest} to #${formattedNewest})`
+  return `Showing transfers between #${formattedOldest} to #${formattedNewest}`
 })
 
 const selectedChain = ref({ label: 'All', value: null as string | null })
@@ -229,7 +229,7 @@ function downloadData() {
           class="flex items-center gap-2 px-2 py-1 text-[12px] font-normal text-[#f5f5f5] bg-[#151515] border border-[#222222] rounded-md hover:bg-[#252525] whitespace-nowrap"
         >
           <IconDownload class="w-4 h-4 text-[#bbbbbb]" />
-          <span class="hidden md:inline">Download Page Data</span>
+          <span class="hidden md:inline">Download</span>
         </button>
       </template>
