+++ components/transaction/CrossChain.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue'
-import { formatJsonPretty } from '~/composables/string'
+import { formatJsonPretty } from '~/composables/useString'
 import { useScreenSize } from '~/composables/useScreenSize'
 import { useFormat } from '~/composables/useFormat'
 import Informational from '~/components/icon/Informational.vue'
