--- composables/blocks.ts
@@ -1,3 +1,0 @@
-export const useBlockMiner = (minerData: string) => JSON.parse(minerData)
-
-export const useBlockCoinbase = (coinbase: string) => JSON.parse(coinbase)
