--- composables/tokens.ts
@@ -1,175 +1,0 @@
-export const unknownNft = {
-  image: '',
-  symbol: 'ukn',
-  id: 'unknown',
-  name: 'unknown',
-}
-
-export const unknownToken = {
-  image: '',
-  symbol: 'ukn',
-  id: 'unknown',
-  name: 'unknown',
-}
-
-export const staticTokens = [
-  {
-    name: "ARKD",
-    symbol: "ARKD",
-    coingeckoId: "",
-    module: "arkade.token",
-    icon: "https://swap.ecko.finance/images/crypto/ark.png",
-  },
-  {
-    id: "kadena",
-    module: "coin",
-    symbol: "KDA",
-    coingeckoId: "",
-    name: "Kadena",
-    icon: "https://swap.ecko.finance/images/crypto/kda-crypto.svg",
-  },
-  {
-    name: "KAYC",
-    symbol: "KAYC",
-    coingeckoId: "",
-    module: "free.KAYC",
-    icon: "https://swap.ecko.finance/images/crypto/kayc.svg",
-  },
-  {
-    symbol: "BKA",
-    coingeckoId: "",
-    name: "Backalley",
-    module: "free.backalley",
-    icon: "https://swap.ecko.finance/images/crypto/bka.svg",
-  },
-  {
-    name: "CRNA",
-    symbol: "CRNA",
-    coingeckoId: "",
-    module: "free.corona-token",
-    icon: "https://swap.ecko.finance/images/crypto/corona.svg",
-  },
-  {
-    name: "Crankk",
-    symbol: "CRKK",
-    coingeckoId: "",
-    module: "free.crankk01",
-    icon: "https://swap.ecko.finance/images/crypto/crankk.png",
-  },
-  {
-    name: "CFLY",
-    symbol: "CFLY",
-    coingeckoId: "",
-    module: "free.cyberfly_token",
-    icon: "https://swap.ecko.finance/images/crypto/cfly.svg",
-  },
-  {
-    name: "DOC",
-    symbol: "DOC",
-    coingeckoId: "",
-    module: "free.docu",
-    icon: "https://swap.ecko.finance/images/crypto/docushield.svg",
-  },
-  {
-    name: "FINUX",
-    symbol: "FINUX",
-    coingeckoId: "",
-    module: "free.finux",
-    icon: "https://swap.ecko.finance/images/crypto/finux.png",
-  },
-  {
-    symbol: "JDE",
-    name: "Jodie",
-    coingeckoId: "",
-    module: "free.jodie-token",
-    icon: "https://swap.ecko.finance/images/crypto/jodie.svg",
-  },
-  {
-    name: "Kapy",
-    symbol: "KAPY",
-    coingeckoId: "",
-    module: "free.kapybara-token",
-    icon: "https://swap.ecko.finance/images/crypto/kapy.svg",
-  },
-  {
-    id: "kishu-ken",
-    symbol: "KISHK",
-    coingeckoId: "",
-    name: "Kishu Ken",
-    module: "free.kishu-ken",
-    icon: "https://swap.ecko.finance/images/crypto/kishk.svg",
-  },
-  {
-    name: "MAGA",
-    symbol: "MAGA",
-    coingeckoId: "",
-    module: "free.maga",
-    icon: "https://swap.ecko.finance/images/crypto/maga.png",
-  },
-  {
-    name: "Hype",
-    symbol: "HYPE",
-    coingeckoId: "",
-    module: "hypercent.prod-hype-coin",
-    icon: "https://swap.ecko.finance/images/crypto/hypercent-crypto.svg",
-  },
-  {
-    id: "kaddex",
-    symbol: "KDX",
-    name: "Kaddex",
-    coingeckoId: "",
-    module: "kaddex.kdx",
-    icon: "https://assets.coingecko.com/coins/images/27325/standard/kdx.png?1696526373",
-  },
-  {
-    id: "kdswap",
-    symbol: "KDS",
-    name: "KDSwap",
-    coingeckoId: "",
-    module: "kdlaunch.kdswap-token",
-    icon: "https://swap.ecko.finance/images/crypto/kds.svg",
-  },
-  {
-    id: "kdlaunch",
-    symbol: "KDL",
-    name: "KDLaunch",
-    coingeckoId: "",
-    module: "kdlaunch.token",
-    icon: "https://swap.ecko.finance/images/crypto/kdl.svg",
-  },
-  {
-    name: "Mok",
-    symbol: "MOK",
-    coingeckoId: "",
-    module: "mok.token",
-    icon: "https://swap.ecko.finance/images/crypto/mok.svg",
-  },
-  {
-    name: "BRO",
-    symbol: "BRO",
-    coingeckoId: "",
-    module: "n_582fed11af00dc626812cd7890bb88e72067f28c.bro",
-    icon: "https://swap.ecko.finance/images/crypto/bro.png",
-  },
-  {
-    name: "zUSD",
-    symbol: "zUSD",
-    coingeckoId: "",
-    module: "n_b742b4e9c600892af545afb408326e82a6c0c6ed.zUSD",
-    icon: "https://swap.ecko.finance/images/crypto/zUSD.svg",
-  },
-  {
-    name: "HERON",
-    symbol: "HERON",
-    coingeckoId: "",
-    module: "n_e309f0fa7cf3a13f93a8da5325cdad32790d2070.heron",
-    icon: "https://swap.ecko.finance/images/crypto/heron.png",
-  },
-  {
-    name: "Flux",
-    symbol: "FLUX",
-    coingeckoId: "",
-    module: "runonflux.flux",
-    icon: "https://swap.ecko.finance/images/crypto/flux-crypto.svg",
-  }
-]
