--- composables/transfer.ts
@@ -1,7 +1,0 @@
-export const useLatestTransfer = (transfers: any[]) => {
-  const nodeLength = transfers.length || 0
-
-  const transferIndex = Math.max(nodeLength - 1, 0)
-
-  return transfers[transferIndex]
-}
