+++ composables/useSearch.ts
@@ -376,70 +376,6 @@ export function useSearch () {
         data.searched = results;
       }
 
-      /*
-       * Disabled: live debounced code search to prevent GraphQL overload.
-       * The following block intentionally commented so code search runs only via Enter flow.
-       *
-       * // Fire code search in background and pop it in first when ready (min 4 chars)
-       * if ((shouldSearchAll || data.filter.value === 'transactions') && value.length >= 4) {
-       *   // mark background pending and flag items
-       *   data._bgPending++;
-       *   if (value === data.query) {
-       *     const curr = { ...(data.searched || results) } as any;
-       *     curr.__bgLoading = true;
-       *     data.searched = curr;
-       *   }
-       *   (async (currentQuery: string) => {
-       *     try {
-       *       const codeResponse: any = await $fetch('/api/graphql', {
-       *         method: 'POST',
-       *         body: {
-       *           query: transactionsByPactCodeQuery,
-       *           variables: {
-       *             pactCode: currentQuery,
-       *             first: 6,
-       *           },
-       *           networkId: selectedNetwork.value?.id,
-       *         },
-       *       });
-       *
-       *       const edges = codeResponse?.data?.transactionsByPactCode?.edges || [];
-       *       if (currentQuery === data.query && edges.length) {
-       *         const codeItems = edges.map((edge: any) => {
-       *           const n = edge.node;
-       *           const resultString = (n?.badResult !== null && n?.badResult !== undefined)
-       *             ? `{"status":"error","badResult":${JSON.stringify(n.badResult)}}`
-       *             : '{"status":"success","badResult":null}';
-       *           return {
-       *             requestkey: n.requestKey,
-       *             chainId: n.chainId,
-       *             height: n.height,
-       *             creationTime: n.creationTime,
-       *             result: resultString,
-       *           };
-       *         });
-       *
-       *         const current = { ...(data.searched || results) } as any;
-       *         current.code = codeItems;
-       *         data.searched = current;
-       *       }
-       *     } catch (error) {
-       *       console.warn('[SEARCH] Code search failed:', error);
-       *     } finally {
-       *       data._bgPending = Math.max(0, (data._bgPending || 0) - 1);
-       *       // if no more background tasks, clear bg flag
-       *       if (data._bgPending === 0) {
-       *         const curr = { ...(data.searched || {}) } as any;
-       *         if (curr.__bgLoading) {
-       *           delete curr.__bgLoading;
-       *           data.searched = curr;
-       *         }
-       *       }
-       *     }
-       *   })(value);
-       * }
-       */
-
       // Fire module search in background if the query matches exact namespace.module heuristic
       const looksLikeModule = (() => {
         const parts = (value || '').split('.')
@@ -634,7 +570,7 @@ export function useSearch () {
     }
 
     // 2. Block Height
-    if (numericRegex.test(searchTerm)) {
+    if (/^\d+$/.test(searchTerm)) {
       const height = parseInt(searchTerm);
       if (height >= 0 && height <= 20000000) {
         try {
@@ -664,7 +600,7 @@ export function useSearch () {
     }
 
     // 3. Transaction Hash - Next Priority (query to verify it exists)
-    if (likelyRequestKeyRegex.test(searchTerm)) {
+    if (/^[A-Za-z0-9\-_]{20,}$/.test(searchTerm)) {
       try {
         const txResponse: any = await $fetch('/api/graphql', {
           method: 'POST',
@@ -1063,3 +999,5 @@ export function useSearch () {
     recordHistory,
   };
 }
+
+
