+++ pages/account/[address].vue
@@ -16,7 +16,7 @@ import AccountNFTTransfers from '~/components/account/AccountNFTTransfers.vue'
 import AccountAssets from '~/components/account/AccountAssets.vue'
 import AccountTokenHoldings from '~/components/account/AccountTokenHoldings.vue'
 import { useAccountBalances } from '~/composables/useAccountBalances'
-import { transformRawBalances } from '~/composables/balances'
+import { transformRawBalances } from '~/composables/useFormat'
 import { useFormat } from '~/composables/useFormat'
 import QrIcon from '~/components/icon/Qr.vue'
 import Tooltip from '~/components/Tooltip.vue'
