+++ pages/blocks/[height]/index.vue
@@ -13,7 +13,7 @@ import { useBlocks } from '~/composables/useBlocks';
 import { useFormat } from '~/composables/useFormat';
 import { useSharedData } from '~/composables/useSharedData';
 import { useScreenSize } from '~/composables/useScreenSize';
-import { exportableToCsv, downloadCSV } from '~/composables/csv';
+import { exportableToCsv, downloadCSV } from '~/composables/useCSV';
 
 definePageMeta({
   layout: 'app',
@@ -49,7 +49,7 @@ const subtitle = computed(() => {
   if (blocks.value.length === 0 || loading.value) {
     return '';
   }
-  return `(Showing all chains for block height #${height.value})`;
+  return `Showing all chains for block height #${height.value}`;
 });
 
 const tableHeaders = [
@@ -135,7 +135,7 @@ function downloadData() {
           class="flex items-center gap-2 px-2 py-1 text-[12px] font-normal text-[#f5f5f5] bg-[#151515] border border-[#222222] rounded-md hover:bg-[#252525] whitespace-nowrap"
         >
           <IconDownload class="w-4 h-4 text-[#bbbbbb]" />
-          <span class="hidden md:inline">Download Page Data</span>
+          <span class="hidden md:inline">Download</span>
         </button>
       </template>
