+++ pages/blocks/index.vue
@@ -18,7 +18,7 @@ import { useBlockCountWss } from '~/composables/useBlockWss';
 import { useFormat } from '~/composables/useFormat';
 import { useSharedData } from '~/composables/useSharedData';
 import { useScreenSize } from '~/composables/useScreenSize';
-import { exportableToCsv, downloadCSV } from '~/composables/csv';
+import { exportableToCsv, downloadCSV } from '~/composables/useCSV';
 
 definePageMeta({
   layout: 'app',
@@ -84,7 +84,7 @@ const subtitle = computed(() => {
   const blockNumbers = filteredBlocks.value.map((b: any) => b.height);
   const oldestBlock = Math.min(...blockNumbers);
   const latestBlock = Math.max(...blockNumbers);
-  return `(Showing blocks between #${oldestBlock} to #${latestBlock})`;
+  return `Showing blocks between #${oldestBlock} to #${latestBlock}`;
 });
 
 const tableHeaders = [
@@ -322,7 +322,7 @@ function downloadData() {
           class="flex items-center gap-2 px-2 py-1 text-[12px] font-normal text-[#f5f5f5] bg-[#151515] border border-[#222222] rounded-md hover:bg-[#252525] whitespace-nowrap"
         >
           <IconDownload class="w-4 h-4 text-[#bbbbbb]" />
-          <span class="hidden md:inline">Download Page Data</span>
+          <span class="hidden md:inline">Download</span>
         </button>
       </template>
