+++ pages/transactions/index.vue
@@ -25,7 +25,7 @@ import { useTransactionCountWss } from '~/composables/useTransactionWss';
 import { useFormat } from '~/composables/useFormat';
 
 // Utils
-import { exportableToCsv, downloadCSV } from '~/composables/csv';
+import { exportableToCsv, downloadCSV } from '~/composables/useCSV';
 
 // Constants
 definePageMeta({
@@ -413,7 +413,7 @@ function downloadData() {
           class="flex items-center gap-2 px-2 py-1 text-[12px] font-normal text-[#f5f5f5] bg-[#151515] border border-[#222222] rounded-md hover:bg-[#252525] whitespace-nowrap"
         >
           <IconDownload class="w-4 h-4 text-[#bbbbbb]" />
-          <span class="hidden md:inline">Download Page Data</span>
+          <span class="hidden md:inline">Download</span>
         </button>
       </template>
